feat(fix): JWT token and upload image size issue fixed

+++ src/app/dashboard/all-blogs/columns.tsx
@@ -11,6 +11,7 @@ import {
 } from "@/components/ui/dropdown-menu"
 import Link from "next/link"
 import { deleteBlog } from "@/services/BlogServices/BlogApi"
+import Swal from "sweetalert2"
 
 // This type is used to define the shape of our data.
 // You can use a Zod schema here if you want.
@@ -53,18 +54,37 @@ export const columns: ColumnDef<IProject>[] = [
 
             // Delete blog
             const handleDeleteBlog = async () => {
-                await deleteBlog(_id)
+                const result = await Swal.fire({
+                    title: "Are you sure woy want to delete this blog?",
+                    text: "You won't be able to revert this!",
+                    icon: "warning",
+                    showCancelButton: true,
+                    confirmButtonColor: "#ff5e14",
+                    cancelButtonColor: "#02245b",
+                    confirmButtonText: "Yes, delete it!"
+                })
+
+                if (result.isConfirmed) {
+                    await deleteBlog(_id)
+                    Swal.fire({
+                        title: "Deleted!",
+                        text: "Blog has been deleted.",
+                        icon: "success",
+                        confirmButtonColor: "#02245b",
+                    });
+                }
+
             }
 
             return (
                 <DropdownMenu>
                     <DropdownMenuTrigger asChild>
-                        <Button variant="ghost" className="h-8 w-8 p-0">
+                        <Button variant="ghost" className="h-8 w-8 p-0 rounded-none cursor-pointer hover:bg-primary">
                             <span className="sr-only">Open menu</span>
                             <MoreHorizontal className="h-4 w-4" />
                         </Button>
                     </DropdownMenuTrigger>
-                    <DropdownMenuContent align="end">
+                    <DropdownMenuContent align="end" className="rounded-none">
                         <DropdownMenuItem><Link href={`/dashboard/projects/${_id}`}>View</Link></DropdownMenuItem>
                         <DropdownMenuItem><Link href={`/dashboard/all-blogs/${_id}`}>Edit</Link></DropdownMenuItem>
                         <DropdownMenuItem onClick={handleDeleteBlog}><span className="cursor-pointer">Delete</span></DropdownMenuItem>
